# docs/swagger.yaml
basePath: /api/v1
definitions:
  httphandler.errorResponse:
    properties:
      message:
        type: string
    type: object
  models.Adv:
    properties:
      color_id:
        minimum: 1
        type: integer
      description:
        maxLength: 150
        minLength: 10
        type: string
      id:
        type: integer
      logo_id:
        minimum: 1
        type: integer
      post:
        maxLength: 100
        minLength: 3
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      tg_id:
        minimum: 1
        type: integer
      title:
        maxLength: 40
        minLength: 3
        type: string
    required:
    - color_id
    - description
    - logo_id
    - post
    - priority
    - tg_id
    - title
    type: object
  models.AdvResponse:
    properties:
      description:
        type: string
      hex:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      post:
        type: string
      priority:
        type: integer
      tg_url:
        type: string
      title:
        type: string
    type: object
  models.AdvUpdate:
    properties:
      color_id:
        minimum: 1
        type: integer
      description:
        maxLength: 150
        minLength: 10
        type: string
      logo_id:
        minimum: 1
        type: integer
      post:
        maxLength: 100
        minLength: 3
        type: string
      priority:
        maximum: 3
        minimum: 1
        type: integer
      tg_id:
        minimum: 1
        type: integer
      title:
        maxLength: 40
        minLength: 3
        type: string
    type: object
  models.Color:
    properties:
      hex:
        maxLength: 7
        minLength: 7
        type: string
      id:
        type: integer
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - hex
    - title
    type: object
  models.ColorResp:
    properties:
      hex:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.ColorUpdate:
    properties:
      hex:
        maxLength: 7
        minLength: 7
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    type: object
  models.ColorsResp:
    properties:
      colors:
        items:
          $ref: '#/definitions/models.ColorResp'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API для сервиса медиа и рекламы
  title: API Documentation
  version: "1.0"
paths:
  /admin/ads:
    post:
      consumes:
      - application/json
      description: Создает новое объявление на основе переданных данных
      parameters:
      - description: Объявление
        in: body
        name: adv
        required: true
        schema:
          $ref: '#/definitions/models.Adv'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание объявления
          schema:
            $ref: '#/definitions/models.AdvResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "409":
          description: Конфликт данных
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Добавить новое объявление
      tags:
      - adv
  /admin/ads/{id}:
    delete:
      description: Удаляет объявление по его id
      parameters:
      - description: id объявления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление
        "404":
          description: Объявление не найдено
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Удалить объявление
      tags:
      - adv
    put:
      consumes:
      - application/json
      description: Обновляет существующее объявление
      parameters:
      - description: id объявления
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленное объявление
        in: body
        name: adv
        required: true
        schema:
          $ref: '#/definitions/models.AdvUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление объявления
          schema:
            $ref: '#/definitions/models.AdvResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "404":
          description: Объявление не найдено
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "409":
          description: Конфликт данных
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Обновить объявление
      tags:
      - adv
  /admin/colors:
    post:
      consumes:
      - application/json
      description: Создает новый цвет
      parameters:
      - description: Цвет
        in: body
        name: color
        required: true
        schema:
          $ref: '#/definitions/models.Color'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание цвета
          schema:
            $ref: '#/definitions/models.ColorResp'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Добавить новый цвет
      tags:
      - colors
  /admin/colors/{id}:
    delete:
      description: Удаляет цвет по его идентификатору
      parameters:
      - description: id цвета
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешное удаление
        "404":
          description: Цвет не найден
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "409":
          description: Конфликт данных
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Удалить цвет
      tags:
      - colors
    put:
      consumes:
      - application/json
      description: Обновляет существующий цвет
      parameters:
      - description: id цвета
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленный цвет
        in: body
        name: color
        required: true
        schema:
          $ref: '#/definitions/models.ColorUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление цвета
          schema:
            $ref: '#/definitions/models.ColorResp'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "404":
          description: Цвет не найден
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Обновить цвет
      tags:
      - colors
...
  /ads:
    get:
      description: Возвращает список объявлений с возможностью фильтрации
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: post
        required: true
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: priority
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Список объявлений
          schema:
            items:
              $ref: '#/definitions/models.AdvResponse'
            type: array
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "404":
          description: Объявления не найдены
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Получить объявления с фильтрацией
      tags:
      - adv
  /colors:
    get:
      description: Возвращает список всех цветов с учетом пагинации
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список цветов
          schema:
            $ref: '#/definitions/models.ColorsResp'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "404":
          description: Цвета не найдены
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Получить список цветов
      tags:
      - colors
...
swagger: "2.0"